[#24] Refactor: LoginModal 컴포넌트에서 모달 창만 Modal 컴포넌트로 분리
